resource-timing/link-sequence-of-events.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Test the sequence of events when reporting link timing.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<script>
  async_test(t => {
    const link = document.createElement('link');
    const delay = 500;
    const src = `./resources/import.sub.css?delay=${delay}`
    const absoluteURL = new URL(src, location.href).toString();
    new PerformanceObserver(t.step_func(() => {
      const entries = performance.getEntriesByType('resource');
      const linkEntry = entries.find(e => e.name.includes('import.sub.css'));
      const importEntry = entries.find(e => e.name.includes('delay-css'));
      if (!linkEntry || !importEntry)
        return;
      console.log({linkEntry, importEntry})
      assert_greater_than_equal(importEntry.startTime + importEntry.duration, linkEntry.startTime, linkEntry.duration + delay);
      t.done();
    })).observe({type: 'resource'});
    link.href = src;
    link.rel = 'stylesheet';
    document.head.appendChild(link);
  }, "test that @imports don't affect link resource timings");
</script>
